+++ internal/vault/cache_storage.go
@@ -36,6 +36,7 @@ const (
 	labelEncrypted       = "encrypted"
 	labelVaultTransitRef = "vaultTransitRef"
 	labelCacheKey        = "cacheKey"
+	labelTokenAccessor   = "vaultTokenAccessor"
 	fieldMACMessage      = "messageMAC"
 	fieldCachedSecret    = "secret"
 
@@ -67,6 +68,9 @@ type ClientCacheStorageStoreRequest struct {
 	Client              Client
 	EncryptionClient    Client
 	EncryptionVaultAuth *secretsv1beta1.VaultAuth
+	// IncludeTokenAccessor is used for testing only, not for production. Exposing
+	// the token accessor in the Secret is a security risk.
+	IncludeTokenAccessor bool
 }
 
 type ClientCacheStoragePruneRequest struct {
@@ -81,13 +85,6 @@ type CachingClientFactoryShutDownRequest struct {
 type ClientCacheStorageRestoreRequest struct {
 	SecretObjKey        ctrlclient.ObjectKey
 	CacheKey            ClientCacheKey
-	DecryptionClient    Client
-	DecryptionVaultAuth *secretsv1beta1.VaultAuth
-	// NoPruneOnError preserves the storage entry on restoration error.
-	NoPruneOnError bool
-}
-
-type ClientCacheStorageRestoreAllRequest struct {
 	DecryptionClient    Client
 	DecryptionVaultAuth *secretsv1beta1.VaultAuth
 	// NoPruneOnError preserves the storage entry on restoration error.
@@ -147,7 +144,6 @@ var _ ClientCacheStorage = (*defaultClientCacheStorage)(nil)
 type ClientCacheStorage interface {
 	Store(context.Context, ctrlclient.Client, ClientCacheStorageStoreRequest) (*corev1.Secret, error)
 	Restore(context.Context, ctrlclient.Client, ClientCacheStorageRestoreRequest) (*clientCacheStorageEntry, error)
-	RestoreAll(context.Context, ctrlclient.Client, ClientCacheStorageRestoreAllRequest) ([]*clientCacheStorageEntry, error)
 	Prune(context.Context, ctrlclient.Client, ClientCacheStoragePruneRequest) (int, error)
 	Purge(context.Context, ctrlclient.Client) error
 	Len(context.Context, ctrlclient.Client) (int, error)
@@ -174,7 +170,7 @@ func (c *defaultClientCacheStorage) getSecret(ctx context.Context, client ctrlcl
 }
 
 func (c *defaultClientCacheStorage) Store(ctx context.Context, client ctrlclient.Client, req ClientCacheStorageStoreRequest) (*corev1.Secret, error) {
-	logger := log.FromContext(ctx)
+	logger := log.FromContext(ctx).V(consts.LogLevelDebug).WithName("clientCacheStorage")
 	var err error
 	defer func() {
 		c.incrementRequestCounter(metrics.OperationStore, err)
@@ -204,8 +200,9 @@ func (c *defaultClientCacheStorage) Store(ctx context.Context, client ctrlclient
 
 	c.mu.Lock()
 	defer c.mu.Unlock()
-	logger.Info("ClientCacheStorage.Store()",
-		"enforceEncryption", c.enforceEncryption)
+	logger.Info("Storing client",
+		"enforceEncryption", c.enforceEncryption,
+		"cacheKey", cacheKey)
 
 	labels := ctrlclient.MatchingLabels{
 		// cacheKey is the key used to access a Client from the ClientCache
@@ -222,6 +219,15 @@ func (c *defaultClientCacheStorage) Store(ctx context.Context, client ctrlclient
 		labelConnectionGeneration: strconv.FormatInt(connObj.Generation, 10),
 		labelProviderUID:          string(credentialProvider.GetUID()),
 		labelProviderNamespace:    credentialProvider.GetNamespace(),
+	}
+
+	// used for testing only, not for production. Exposing the token accessor in the
+	// Secret is a security risk.
+	if req.IncludeTokenAccessor {
+		accessor, err := req.Client.GetTokenSecret().TokenAccessor()
+		if err == nil {
+			labels[labelTokenAccessor] = accessor
+		}
 	}
 	s := &corev1.Secret{
 		// we always store Clients in an Immutable secret as an anti-tampering mitigation.
@@ -292,9 +298,16 @@ func (c *defaultClientCacheStorage) Store(ctx context.Context, client ctrlclient
 			}
 		} else {
 			err = e
+			logger.Error(err, "Failed to store client",
+				"enforceEncryption", c.enforceEncryption,
+				"cacheKey", cacheKey, "secret", ctrlclient.ObjectKeyFromObject(s))
 			return nil, err
 		}
 	}
+
+	logger.Info("Stored client",
+		"enforceEncryption", c.enforceEncryption,
+		"cacheKey", cacheKey, "secret", ctrlclient.ObjectKeyFromObject(s))
 
 	return s, nil
 }
@@ -495,44 +508,6 @@ func (c *defaultClientCacheStorage) Purge(ctx context.Context, client ctrlclient
 
 	err = client.DeleteAllOf(ctx, &corev1.Secret{}, c.deleteAllOfOptions()...)
 	return err
-}
-
-func (c *defaultClientCacheStorage) RestoreAll(ctx context.Context, client ctrlclient.Client, req ClientCacheStorageRestoreAllRequest) ([]*clientCacheStorageEntry, error) {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-
-	var errs error
-	defer func() {
-		c.incrementRequestCounter(metrics.OperationRestoreAll, errs)
-	}()
-
-	found, err := c.listSecrets(ctx, client, c.listOptions()...)
-	if err != nil {
-		errs = errors.Join(err)
-		return nil, errs
-	}
-
-	var result []*clientCacheStorageEntry
-	for _, s := range found {
-		cacheKey := ClientCacheKey(s.Labels[labelCacheKey])
-		req := ClientCacheStorageRestoreRequest{
-			SecretObjKey:        ctrlclient.ObjectKeyFromObject(&s),
-			CacheKey:            cacheKey,
-			DecryptionClient:    req.DecryptionClient,
-			DecryptionVaultAuth: req.DecryptionVaultAuth,
-			NoPruneOnError:      req.NoPruneOnError,
-		}
-
-		entry, err := c.restore(ctx, client, req, &s)
-		if err != nil {
-			errs = errors.Join(errs, err)
-			continue
-		}
-
-		result = append(result, entry)
-	}
-
-	return result, errs
 }
 
 func (c *defaultClientCacheStorage) validateSecretMAC(req ClientCacheStorageRestoreRequest, s *corev1.Secret) error {