+++ test/integration/operator/terraform/main.tf
@@ -107,6 +107,6 @@ module "vso-helm" {
   }
   manager_extra_args = [
     "-min-refresh-after-hvsa=3s",
-    "-zap-log-level=5"
+    "-zap-log-level=6"
   ]
 }